refactor: rename client.py to client_utils.py
